requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# AI & ML
anthropic==0.34.0
google-generativeai==0.8.3
openai==1.51.0
pdf2image==1.17.0
pillow==10.4.0

# Validation
pydantic==2.9.1
pydantic-settings==2.5.2
email-validator==2.2.0

# CORS & Middleware
python-dotenv==1.0.1

# Utilities
uuid==1.30